# Conflicts:
#	src/main/java/co/acelerati/planetexpress/infraestructure/http/rest/controller/StockRestController.java

+++ src/main/resources/scripts/schema_planet_express.sql
@@ -5,8 +5,8 @@ CREATE SCHEMA public AUTHORIZATION postgres;
 
 -- DROP TABLE public.stock;
 CREATE TABLE public.stock (
-	product_id 					NUMERIC(1000) NOT NULL,
-	quantity 					NUMERIC(1000) NOT NULL,
+	product_id 					INTEGER NOT NULL,
+	quantity 					INTEGER NOT NULL,
 	current_price 				NUMERIC(16, 2) NOT NULL,
 
 	CONSTRAINT stock_pk 		PRIMARY KEY(product_id)
@@ -15,7 +15,7 @@ CREATE TABLE public.stock (
 -- DROP TABLE public.supply;
 CREATE TABLE public.supply (
 	supply_id					UUID,
-	supplier_id					NUMERIC(1000) NOT NULL,
+	supplier_id					INTEGER NOT NULL,
 	date_supply					DATE NOT NULL,
 
 	CONSTRAINT supply_pk 		PRIMARY KEY(supply_id)
@@ -25,8 +25,8 @@ CREATE TABLE public.supply (
 CREATE TABLE public.supply_stock (
 	supply_stock_id 			UUID,
 	supply_id 					UUID NOT NULL,
-	stock_id 					NUMERIC(1000) NOT NULL,
-	quantity 					NUMERIC(1000) NOT NULL,
+	stock_id 					INTEGER NOT NULL,
+	quantity 					INTEGER NOT NULL,
 	supply_price 				NUMERIC(16, 2) NULL,
 
 	CONSTRAINT supply_stock_pk	PRIMARY KEY(supply_stock_id),
@@ -37,7 +37,7 @@ CREATE TABLE public.supply_stock (
 -- DROP TABLE public.shopping_cart;
 CREATE TABLE public.shopping_cart (
 	shopping_cart_id 				UUID,
-	user_id 						NUMERIC(1000) NOT NULL,
+	user_id 						INTEGER NOT NULL,
 	last_update 					DATE NULL,
 
 	CONSTRAINT shopping_cart_pk 	PRIMARY KEY(shopping_cart_id)
@@ -46,9 +46,9 @@ CREATE TABLE public.shopping_cart (
 -- DROP TABLE public.shopping_cart_stock;
 CREATE TABLE public.shopping_cart_stock (
 	shopping_cart_stock_id				UUID,
-	stock_id 							NUMERIC(1000) NOT NULL,
+	stock_id 							INTEGER NOT NULL,
 	shopping_cart_id					UUID NOT NULL,
-	quantity 							NUMERIC(1000) NOT NULL,
+	quantity 							INTEGER NOT NULL,
 
 	CONSTRAINT shopping_cart_stock_pk	PRIMARY KEY(shopping_cart_stock_id),
 	CONSTRAINT stock_fk 				FOREIGN KEY(stock_id) REFERENCES public.stock(product_id),
@@ -58,8 +58,8 @@ CREATE TABLE public.shopping_cart_stock (
 -- DROP TABLE public.sale;
 CREATE TABLE public.sale (
 	sale_id 					UUID,
-	client_id 					NUMERIC(1000) NOT NULL,
-	salesman_id 				NUMERIC(1000) NOT NULL,
+	client_id 					INTEGER NOT NULL,
+	salesman_id 				INTEGER NOT NULL,
 	date_sale 					DATE NOT NULL,
 	state 						VARCHAR(12) NULL,
 	
@@ -70,7 +70,7 @@ CREATE TABLE public.sale (
 CREATE TABLE public.sale_stock (
 	sale_stock_id 				UUID,
 	sale_id 					UUID NOT NULL,
-	stock_id 					NUMERIC(1000) NOT NULL,
+	stock_id 					INTEGER NOT NULL,
 	quantity 					NUMERIC(16, 2) NOT NULL,
 	current_price				NUMERIC(16, 2) NOT NULL,
